move should resolve label name tokens to label values from query response test to prometheus-plugin utils.test.ts
